requirements.txt
numpy>=1.21.0
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.12.0
opencv-python>=4.6.0
pandas>=1.4.0
scipy>=1.8.0
pyserial>=3.5
